Added the base files for the initial version.
Pokegenesys Monster Data contains the stat block for every official pokemon and official variant present so far.
Moves Module contains the building blocks necessary to create moves or attacks within the system. It is based upon the Signature Spell section of the Genesys handbook.
Catching Module contains the basic data for catching a pokemon with a ball. Several balls are listed, but the expansion of this module is planned.
Berries Module contains data for every official berry variant along with their effects. Some effects have been altered from their original versions to provide more versatility within the system.
Emotion Engine Module contains a method to expand on the motivations section of the Genesys handbook. This section is a more robust method of determining a creature's actions based on stimuli. It is especially geared towards simulating the pokemon's side of a relationship between themselves and their trainer. It is entirely optional.

Current Roadmap:
1: Testing
2: Adding mainstay items such as Potion and Full Heal
3: Clarifying evolution language.
4: Clarifying XP gain and spending.
5: Creating a Trainer talent that unlocks and expands the leadership skill for checks involving leading a pokemon
6: Expanding the Catching module to include additional ball types.
7: Adding a crafting module to include Ball and Potion crafting
8: Determining cost conversions between standard Pokemon sources and those used by the standard Genesys loadout.
9: Putting everything together in a pdf or booklet that can be handed to players.
10: Creating a searchable database that allows for encounter building and NPC generation.
11: Creation of an Ability Module that adds the Pokemon Abilities from the main series. May look into converting heroic abilities.
12: Assigning biome data to the Monster database for use with encounter building.
13: Converting/creating a compatible pokemon-themed campaign for use with the system. I actually have one built, but a new one may be needed for a more authentic adventure.
14: Creation of a Mystery Dungeon module. A set of rules for when the PCs are pokemon themselves.